# improve performance and memory footprint of inserts