@Serabe 2017-08-02
By moving away from squirrel for updates and deletes, you mean for specific record updates and deletes, not for batch ones, don't you?

Building queries can get really complicated really quick for non-trivial cases.