@erizocosmico 2017-08-02
There are no batch updates/deletes.

Yes, querying can get ugly. But the only complex things we're doing are joins and counts. Not like we have to implement a full SQL builder. It's not like we're gonna do this now, but it would be nice for the future to move to a non-immutable query builder, because this one does a lot of allocs.